аÑfix csp
