+++ packages/server/src/index.ts
@@ -15,7 +15,7 @@ import cors from 'cors'
 import staticMiddleware from './modules/middlewares/static.middleware'
 import ssrMiddleware from './modules/middlewares/ssr.middleware'
 import proxyMiddleware from './modules/middlewares/proxy.middleware'
-import cspMiddleware from './modules/middlewares/cspMiddleware'
+import cspMiddleware from './modules/middlewares/csp.middleware'
 
 import routerForum from './modules/forum/forum.router'
 import routerTopic from './modules/topic/topic.router'
@@ -53,7 +53,7 @@ async function start() {
   // * MIDDLEWARES
   app.use(cors())
   app.use(cookieParser())
-  // app.use(cspMiddleware())
+  app.use(cspMiddleware())
 
   app.use('/assets', staticMiddleware())
   app.use('/api/v2', proxyMiddleware())
